Removed the MkvWriter trait
Instead, the mkv_writer() method used by Segment is put directly on
Writer instead. Note that this change also means we can't use that
MkvWriter trait as a bound on Segment::new(), and we can't use Writer
instead because it's a struct, not a trait. So, Segment now contains a
Writer<W> instead of just a W, meaning that what was once written
Segment<Writer<File>> is now just Segment<File>.